requirements.txt
requests==2.30.0
python-dotenv==1.0.0
python-telegram-bot==13
google-cloud-dialogflow==2.23.3
google-cloud==0.34.0
google-cloud-api-keys==0.5.3
